+++ src/app/Credits.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Credits = () => (
+  <p>Credit for the favicon goes to <a href="https://www.flaticon.com/free-icon/dollar_550638">flaticon</a></p>
+);
+
+export default Credits;